requirements.txt
Django==2.0.7
django-bootstrap3==10.0.1
django-crispy-forms==1.7.2
django-filter==2.0.0
django-tables2==1.21.2
pytz==2018.5
